pyproject.toml
[project]
name = "assignment-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "scalar-fastapi>=1.6.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
